requirements/base.txt
# General
Django==1.9.12


# Celery
celery==4.1.0
# amqp==2.2.2
# billiard==3.5.0.3
kombu==4.1.0
# pytz==2018.3
# vine==1.1.4


# Django Rest Framework
djangorestframework==3.7.3


# Cors Headers
django-cors-headers==2.1.0


# Editor
django-wysiwyg-redactor==0.4.9.1


# Map
django-geoposition==0.3.0


# Image
easy-thumbnails==2.5
Pillow==4.0.0
# olefile==0.44


# Sortable
django-admin-sortable==2.0.21


# Twitter
python-twitter==3.2.1
# future==0.16.0
# oauthlib==2.0.1
# requests-oauthlib==0.8.0
twitter-text-python==1.1.0


# Captcha
django-recaptcha2==1.0.0


# Disqus
django-disqus==0.5
# mock==2.0.0
# pbr==3.0.0


# Import Export
django-import-export==1.0.0
# diff-match-patch==20121119
# et-xmlfile==1.0.1
# jdcal==1.3
# odfpy==1.3.6
# openpyxl==2.5.1
# pyyaml==3.12
# tablib==0.12.1
# unicodecsv==0.14.1
# xlrd==1.1.0
# xlwt-1.3.0
